Aula 04

+++ src/controllers/games-controller.ts
@@ -11,7 +11,7 @@ export async function getGames(req: Request, res: Response) {
 
 export async function getGameById(req: Request, res: Response) {
   const id = Number(req.params.id);
-  if (!id || isNaN(id)) throw invalidIdError();
+  if (!id || isNaN(id) || id < 1) throw invalidIdError();
 
   const game = await gamesService.getGameById(id);
   res.status(httpStatus.OK).send(game);
@@ -27,7 +27,7 @@ export async function createGame(req: Request, res: Response) {
 export async function editGame(req: Request, res: Response) {
   const { name } = req.body as Omit<Game, 'id'>;
   const id = Number(req.params.id);
-  if (!id || isNaN(id)) throw invalidIdError();
+  if (!id || isNaN(id) || id < 1) throw invalidIdError();
 
   const game = await gamesService.editGame(id, name);
   res.status(httpStatus.OK).send(game);
@@ -35,8 +35,8 @@ export async function editGame(req: Request, res: Response) {
 
 export async function deleteGame(req: Request, res: Response) {
   const id = Number(req.params.id);
-  if (!id || isNaN(id)) throw invalidIdError();
+  if (!id || isNaN(id) || id < 1) throw invalidIdError();
 
-  await gamesService.deleteGame(id);
-  res.sendStatus(httpStatus.OK);
+  const game = await gamesService.deleteGame(id);
+  res.status(httpStatus.OK).send(game);
 }
